requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==24.3.0
beartype==0.19.0
black==24.10.0
cffi==1.15.1  
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
colorspacious==1.1.2
contourpy==1.3.1
cycler==0.12.1
dacite==1.8.1
decorator==4.0.2
e3nn==0.5.5
einops==0.8.0
filelock==3.12.2  
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
hf-xet==1.1.7  
huggingface-hub==0.30.2
hydra-core==1.3.2
idna==3.4
imageio==2.36.1
imageio-ffmpeg==0.5.1
iopath==0.1.10
jaxtyping==0.2.36
Jinja2==3.1.2 
kiwisolver==1.4.7
lazy_loader==0.4
lightning==2.4.0
lightning-utilities==0.11.9
lpips==0.1.4
MarkupSafe==2.1.1
matplotlib==3.10.0
mkl-service==2.4.0
mkl_fft
mkl_random
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==1.0.0
networkx==3.1
numpy==2.2.0
omegaconf==2.3.0
opencv-python-headless==4.10.0.84
opt-einsum-fx==0.1.4
opt_einsum==3.4.0
packaging==24.0
pathspec==0.12.1
pillow==10.4.0
platformdirs==3.10.0
plyfile==1.1
portalocker==3.0.0
proglog==0.1.10
propcache==0.2.1
protobuf==5.29.1
pycparser==2.21
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
roma==1.5.1
ruff==0.8.3
safetensors==0.4.5
scikit-image==0.25.0
scikit-video==1.1.11
scipy==1.14.1
sentry-sdk==2.19.2
six==1.16.0
smmap==5.0.1
svg.py==1.5.0
sympy==1.13.1
tabulate==0.9.0
timm==1.0.12
tqdm
triton==3.1.0
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.0.3
wadler_lindig==0.1.7
wandb==0.19.2
yarl==1.18.3

diff_gauss_pose @ git+https://github.com/slothfulxtx/diff-gaussian-rasterization.git@pose#egg=diff-gauss-pose
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@055ab3a2e3e611dff66fa82f632e62a315f3b5e7